% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_harmonicDist}
\alias{C_harmonicDist}
\title{Harmomic mean distance}
\usage{
C_harmonicDist(data, subset, S_inv)
}
\description{
Calculates distance between observations as the harmomic mean distance of every point from a selected subset of points (subset defaults to all points).
}
\keyword{internal}

