% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_harmonicDist_partial}
\alias{C_harmonicDist_partial}
\title{Harmomic mean distance using chunk of data to enable the use of a progress bar for large datasets}
\usage{
C_harmonicDist_partial(data, S_inv, i_start, i_end)
}
\description{
Equivalent to C_harmonicDist, but only performs calculation on chunk of data, allowing calculation to be broken into sections and thus tracked by a progress bar.
}
\keyword{internal}

