% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_kernelDeviance}
\alias{C_kernelDeviance}
\title{Overall deviance for given kernel bandwidth}
\usage{
C_kernelDeviance(data, subset, sigma2, S_inv)
}
\description{
Calculates overall deviance (-2*log-likelihood) for given kernel bandwidth. Likelihood of point i is equal to kernel density from all other points in subset, and likelihood of overall data is product over i.
}
\keyword{internal}

