% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_kernelDeviance_partial}
\alias{C_kernelDeviance_partial}
\title{overall deviance for given kernel bandwidth using chunk of data to enable the use of a progress bar for large datasets}
\usage{
C_kernelDeviance_partial(data, sigma2, S_inv, i_start, i_end)
}
\description{
Equivalent to C_kernelDeviance, but only performs calculation on chunk of data, allowing calculation to be broken into sections and thus tracked by a progress bar.
}
\keyword{internal}

