% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_kernelDist}
\alias{C_kernelDist}
\title{kernel density}
\usage{
C_kernelDist(data, subset, sigma2, S_inv)
}
\description{
calculates kernel density of all points from a subset of points.
}
\keyword{internal}

