% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_neighborDist}
\alias{C_neighborDist}
\title{Nearest neighbor distance}
\usage{
C_neighborDist(data, subset, S_inv)
}
\description{
Calculates nearest neighbor distance between all points and a subset of points.
}
\keyword{internal}

