% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{DCMS}
\alias{DCMS}
\title{De-correlated Composite of Multiple Signals (DCMS)}
\usage{
DCMS(dfv, column.nums = 1:ncol(dfv), subset = 1:nrow(dfv), S = NULL, dfp,
  column.nums.p = 1:ncol(dfp))
}
\arguments{
\item{dfv}{a data frame containing observations in rows and raw statistics in columns.}

\item{column.nums}{indexes the columns of \code{dfv} that contain raw statistics.}

\item{subset}{index the rows of the data frame that will be used to calculate the covariance matrix S (unless specified manually).}

\item{S}{the covariance matrix used to account for correlation between observations in the DCMS calculation. Leave as NULL to use the ordinary covariance matrix calculated using \code{cov(dfv[subset,column.nums],use="pairwise.complete.obs")}.}

\item{dfp}{a data frame containing observations in rows and p-values in columns.}

\item{column.nums.p}{indexes the columns of \code{dfp} that contain p-values.}
}
\description{
Calculates the DCMS for each row (locus, SNP) in the data frame. Data are subset prior to calculating distances (see details).
}
\details{
The selected columns of the \code{dfv} data frame (i.e. the columns specified by \code{column.nums}) should contain the raw test statistics, while the selected columns of the \code{dfp} data frame (i.e. the columns specified by \code{column.nums.p}) should contain the corresponding p-values. If the same data frame contains both raw statistics and p-values then this should be passed in as both \code{dfv} and \code{dfp}, with only the selected columns changing. The covariance matrix used in the DCMS calculation can be specified directly through the argument S, or if S=NULL then this matrix is calculated directly from selected rows and columns of \code{dfv}.
}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

