% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/foo1.R
\name{foo1}
\alias{foo1}
\title{This is foo1.
This function provides vital tools to do basic arithmetic.}
\usage{
foo1(x, y)
}
\arguments{
\item{x}{a number or integer.}

\item{y}{a number or integer.}
}
\description{
This is foo1.
This function provides vital tools to do basic arithmetic.
}
\examples{
## try out our super useful function
foo1(1, 5)
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

