% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanGWAS}
\alias{HumanGWAS}
\title{HumanGWAS output for 17,624 SNPs.}
\format{A data frame with 17,624 rows and 9 variables in the columns.}
\source{
Unpublished human GWAS analysis.
}
\usage{
HumanGWAS
}
\arguments{
\item{Chr}{A factor containing the chromosome of each individual site.}
}
\description{
This dataset contains example output from a human Genome-Wide Association Study (GWAS).
}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}

Liuyang Wang \email{wallacewly@gmail.com}
}
\keyword{datasets}

