% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/KernelDensityML.R
\name{KernelDensityML}
\alias{KernelDensityML}
\title{R Verity
 March 18, 2015
 Kernel density with maximum likelihood bandwidth selection}
\usage{
KernelDensityML(dfv, column.nums)
}
\arguments{
\item{dfv}{is a dataframe containing the observations in rows and the statistics in columns}

\item{column.nums}{is the column numbers of the dataframe containing the statistics used to calculate kernel density distance}
}
\description{
R Verity
 March 18, 2015
 Kernel density with maximum likelihood bandwidth selection
}
\examples{
##TO DO
}
\author{
R Verity
}

