% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{kernelDist}
\alias{kernelDist}
\title{Kernel Density Distance}
\usage{
kernelDist(dfv, column.nums = 1:ncol(dfv), bandwidth = "default",
  S = NULL)
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns.}

\item{column.nums}{indexes the columns of the data frame that contain the statistics used to
calculate kernel distance (all other columns are ignored).}

\item{bandwidth}{standard deviation of the normal kernel in each dimension. Can be a numerical value, or can be set to 'default', in which case Silverman's rule is used to select the bandwidth.}

\item{S}{the covariance matrix that the bandwidth is multiplied by. Leave as NULL to use the ordinary covariance matrix calculated on the data.}
}
\description{
Calculates kernel density of all points from all others in multivariate space. Returns -2*log(density)
as a distance measure.
}
\details{
Takes a matrix or data frame as input, with observations in rows and statistics in columns.
Values are first normalized in each dimension by subtracting from the
mean and dividing by the standard deviation before calculating kernel density distance. Assumes a
multivariate normal kernel with the same user-defined bandwidth in all dimensions (after normalization).

Note that this method cannot handle NA values.
}
\examples{

# create a data frame of observations
df <- data.frame(x=rnorm(100),y=rnorm(100))

# calculate kernel density distances
distances <- kernelDist(df)

# use this distance to look for outliers
Q95 <- quantile(distances, 0.95)
which(distances>Q95)

}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

