% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mahalonobis.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{Mahalanobis}
\usage{
Mahalanobis(dfv, column.nums = 1:ncol(dfv))
}
\description{
Calculates the Mahalanobis distance for each row (locus, SNP) in the data frame. The input data frame can handle some missing data, as long as a covariance matrix can still be computed.
 @param dfv is a data frame containing observations in rows and statistics in columns
 @param column.nums is the column numbers of the data frame containing the statistics used to calculate Mahalanobis distance (other columns are ignored)
}
\examples{

# (STILL NEEDS DOING)

}
\author{
KE Lotterhos \email{k.lotterhos@neu.edu}
}

