% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/startServer.R
\name{MANIPULATE}
\alias{MANIPULATE}
\title{This is MANIPULATE.
This function calls the shiny app.}
\usage{
MANIPULATE()
}
\description{
This is MANIPULATE.
This function calls the shiny app.
}
\author{
Caitlin Collins \email{caitlin.collins12@imperial.ac.uk}
}

