% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meeting-minutes.R
\name{meeting_minutes}
\alias{meeting_minutes}
\title{Meeting Minutes}
\usage{
meeting_minutes(css = "deafult")
}
\arguments{
\item{css}{\strong{character} name of css file(s)}
}
\value{
An R markdown output format
}
\description{
Based on \code{pagedown::html_paged()}.
}
