% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINOTAUR.R
\name{MINOTAUR}
\alias{MINOTAUR}
\title{Run the MINOTAUR app!}
\usage{
MINOTAUR(x = NULL)
}
\arguments{
\item{x}{A dummy variable.}
}
\description{
Running this function will open the MINOTAUR app, a
shiny application that can be used to explore data, compare variables, 
and identify outliers in multivariate and univariate space.
}
\examples{

## run function
MINOTAUR()

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

