% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NonParametricInverse}
\alias{NonParametricInverse}
\title{Example Dataset: Non-Parametric Inverse}
\format{A data frame with 1,002 observations in the rows
 and 2 variables in the columns.}
\source{
Simulated non-parametric dataset.
}
\usage{
NonParametricInverse
}
\description{
This is a simple non-parametric two-variable dataset
with an inverse relationship between variables.
}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}

Caitlin Collins \email{caitlin.collins12@imperial.ac.uk}
}
\references{
Robert Verity*, Caitlin Collins*,
Daren C. Card, Sara M. Schaal, Liuyang Wang, & Katie E. Lotterhos.
MINOTAUR: an R package for visualizing and
calculating multivariate outliers in genomic datasets.
In Review.
}
\keyword{datasets}

