% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NonParametricMultimodal}
\alias{NonParametricMultimodal}
\title{Example Dataset: Nonparametric Multimodal}
\format{A data frame with 1,000 observations in the rows
 and 2 variables in the columns.}
\source{
Simulated nonparametric dataset; unpublished.
}
\usage{
NonParametricMultimodal
}
\description{
This is a simple nonparametric two-variable dataset
that is multimodal.
}
\details{
\itemize{
\item \code{x} \itemize{\item A simulated numeric variable.}
\item \code{y} \itemize{\item A simulated numeric variable.}
}
}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}

Caitlin Collins \email{caitlin.collins12@imperial.ac.uk}
}
\references{
Robert Verity*, Caitlin Collins*,
Daren C. Card, Sara M. Schaal, Liuyang Wang, & Katie E. Lotterhos.
MINOTAUR: an R package for visualizing and
calculating multivariate outliers in genomic datasets.
In Review.
}
\keyword{datasets}

