% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{stat_to_pvalue}
\alias{stat_to_pvalue}
\title{Convert raw statistics to p-values using one of a number of possible methods.}
\usage{
stat_to_pvalue(dfv, column.nums = 1:ncol(dfv), subset = 1:nrow(dfv),
  two.tailed = TRUE, right.tailed = TRUE)
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns.}

\item{column.nums}{indexes the columns of the data frame that will be used to
calculate p-values (all other columns are ignored).}

\item{subset}{index the rows of the data frame that are known contain values from the null distribution. Use all rows if no such information is available.}

\item{two.tailed}{whether p-value calculations should be two tailed.}

\item{right.tailed}{if using one-tailed test, whether that tail should be in the positive direction.}
}
\description{
Text
}
\details{
Detailed text
}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

