% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TwoRefSim}
\alias{TwoRefSim}
\title{Example Dataset: Simulated Expansion from Two Refugia}
\format{A data frame with 10,000 Single-Nucleotide Polymorphisms (SNPs)
in the rows and 15 variables in the columns.}
\source{
Simulated population genetic dataset.
}
\usage{
TwoRefSim
}
\description{
This dataset contains population genetic data simulated
under a model of expansion from two refugia.
}
\details{
\itemize{
\item \code{SNPnames} \itemize{\item A factor containing the unique names of each SNP locus.}
\item \code{Corr.all} \itemize{\item A ...}
\item \code{FST.All} \itemize{\item A ...}
\item \code{s_high} \itemize{\item A ...}
\item \code{SNPIncluded} \itemize{\item A ...}
\item \code{He.LS} \itemize{\item A ...}
\item \code{He_samp} \itemize{\item A ...}
\item \code{log.bf} \itemize{\item A ...}
\item \code{rho} \itemize{\item A ...}
\item \code{xtx} \itemize{\item A ...}
\item \code{TW.Zscore} \itemize{\item A ...}
\item \code{Md} \itemize{\item A ...}
\item \code{Hd} \itemize{\item A ...}
\item \code{Kd} \itemize{\item A ...}
\item \code{Nd} \itemize{\item A ...}
}
}
\author{
Kathleen Lotterhos \email{k.lotterhos@neu.edu}

Caitlin Collins \email{caitlin.collins12@imperial.ac.uk}
}
\references{
Lotterhos KE, Whitlock MC (2015)
The relative power of genome scans to detect local adaptation
depends on sampling design and statistical method.
Molecular Ecology, 24, 1031–1046.

Robert Verity*, Caitlin Collins*,
Daren C. Card, Sara M. Schaal, Liuyang Wang, & Katie E. Lotterhos.
MINOTAUR: an R package for visualizing and
calculating multivariate outliers in genomic datasets.
In Review.
}
\keyword{datasets}

