% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{CSS}
\alias{CSS}
\title{Composite Selection Signal}
\usage{
CSS(dfv, column.nums = 1:ncol(dfv), subset = 1:nrow(dfv),
  two.tailed = rep(TRUE, length(column.nums)), right.tailed = rep(FALSE,
  length(column.nums)))
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns.}

\item{column.nums}{indexes the columns of the data frame that will be used to
calculate CSS (all other columns are ignored). Only columns containing raw statistics should be used in the CSS calculation, and not those containing p-values generated from raw statistics.}

\item{two.tailed}{whether p-value calculations should be two tailed.}

\item{right.tailed}{if using one-tailed test, whether that tail should be in the positive direction.}
}
\description{
Calculates the CSS for each row (locus, SNP) in the data frame. Data are subset prior to calculating distances (see details).
}
\details{
Calculates CSS based on the method described in Randhawa et al (2014). Selected columns of \code{dfv} are first converted to fractional ranks (see \code{stat_to_pvalue}). Fractional ranks are then converted to z-scores using the inverse cumulative normal transformation. The mean z-score is then taken over variables, and converted to a p-value based on the appropriate normal distribution. Finally, CSS is defined as -log(p-value) in base 10.

Text.
}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}
\references{
Randhawa, Imtiaz Ahmed Sajid, et al. "Composite selection signals can localize the trait specific genomic regions in multi-breed populations of cattle and sheep." BMC genetics 15.1 (2014): 1.
}

