% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{CSS}
\alias{CSS}
\title{Composite Selection Signal}
\usage{
CSS(dfv, column.nums = 1:ncol(dfv), subset = 1:nrow(dfv),
  two.tailed = rep(TRUE, length(column.nums)), right.tailed = rep(FALSE,
  length(column.nums)))
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns.}

\item{column.nums}{indexes the columns of the data frame that will be used to
calculate CSS (all other columns are ignored). Only columns containing raw statistics should be used in the CSS calculation, and not those containing p-values generated from raw statistics.}

\item{two.tailed}{whether p-value calculations should be two tailed.}

\item{right.tailed}{if using one-tailed test, whether that tail should be in the positive direction.}
}
\description{
Calculates the CSS for each row (locus, SNP) in the data frame. Data are subset prior to calculating distances (see details).
}
\details{
Detailed text
}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

