% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{harmonicDist}
\alias{harmonicDist}
\title{Harmonic Mean Distance}
\usage{
harmonicDist(dfv, column.nums = 1:ncol(dfv))
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns}

\item{column.nums}{indexes the columns of the data frame that contain the statistics used to
calculate Mahalanobis distance (all other columns are ignored)}
}
\description{
Calculates harmonic mean
distance of all points from all others.
}
\details{
Takes a matrix or data frame as input, with observations in rows and statistics in columns.
Values are first standardized in each dimension by subtracting from the
mean and dividing by the standard deviation before calculating harmonic mean distances.

Note that dfv cannot contain any NA values.
}
\examples{

# create a data frame of observations
df <- data.frame(x=rnorm(100),y=rnorm(100))

# calculate harmonic mean distances
distances <- harmonicDist(df)

# use this distance to look for outliers
Q95 <- quantile(distances, 0.95)
which(distances>Q95)

}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

