% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HumanGWAS}
\alias{HumanGWAS}
\title{Example Dataset: Human GWAS}
\format{A data frame with 17,624 Single-Nucleotide Polymorphisms (SNPs)
 in the rows and 9 variables in the columns.}
\source{
Unpublished human GWAS analysis.
}
\usage{
HumanGWAS
}
\description{
This dataset contains an example of output returned
from a human Genome-Wide Association Study (GWAS).
}
\details{
\itemize{
\item \code{Chr} \itemize{\item A factor containing the chromosome on which each SNP locus is located.}
\item \code{SNP} \itemize{\item A ...}
\item \code{BP} \itemize{\item A ...}
\item \code{Trait1_Beta} \itemize{\item A ...}
\item \code{Trait1_P} \itemize{\item A ...}
\item \code{Trait2_Beta} \itemize{\item A ...}
\item \code{Trait2_P} \itemize{\item A ...}
\item \code{Trait3_Beta} \itemize{\item A ...}
\item \code{Trait3_P} \itemize{\item A ...}
}
}
\author{
Liuyang Wang \email{wallacewly@gmail.com}

Caitlin Collins \email{caitlin.collins12@imperial.ac.uk}
}
\references{
Robert Verity*, Caitlin Collins*,
Daren C. Card, Sara M. Schaal, Liuyang Wang, & Katie E. Lotterhos.
MINOTAUR: an R package for visualizing and
calculating multivariate outliers in genomic datasets.
In Review.
}
\keyword{datasets}

