% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{kernelDist}
\alias{kernelDist}
\title{Kernel Density Distance}
\usage{
kernelDist(dfv, column.nums = 1:ncol(dfv), bandwidth = "default")
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns}

\item{column.nums}{indexes the columns of the data frame that contain the statistics used to
calculate Mahalanobis distance (all other columns are ignored)}

\item{bandwidth}{standard deviation of the normal kernel in each dimension}
}
\description{
Calculates kernel density of all points from all others in multivariate space. Returns -2*log(density)
as a distance measure.
}
\details{
Takes a matrix or data frame as input, with observations in rows and statistics in columns.
Values are first standardized in each dimension by subtracting from the
mean and dividing by the standard deviation before calculating kernel density distance. Assumes a
multivariate normal kernel with the same user-defined bandwidth in all dimensions (but note that
the bandwidth is applied after normalization).

Note that dfv cannot contain any NA values.
}
\examples{

# create a data frame of observations
df <- data.frame(x=rnorm(100),y=rnorm(100))

# calculate kernel density distances
distances <- harmonicDist(df)

# use this distance to look for outliers
Q95 <- quantile(distances, 0.95)
which(distances>Q95)

}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

