% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{kernelLogLike}
\alias{kernelLogLike}
\title{Kernel Density Log-likelihood}
\usage{
kernelLogLike(dfv, column.nums = 1:ncol(dfv), bandwidth = 1:5,
  reportProgress = FALSE)
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns}

\item{column.nums}{indexes the columns of the data frame that contain the statistics used to
calculate Mahalanobis distance (all other columns are ignored)}

\item{bandwidth}{a vector containing the range of bandwidths to be explored}
}
\description{
Calculates log-likelihood of data under kernel density model for a range of bandwidths.
Can be used to estimate the optimal (maximum likelihood) bandwith to use in functions
such as kernelDist().
}
\details{
Uses same input and model structure as kernelDist(). Calculates the log-likelihood using the
leave-one-out method, wherein the likelihood of each point is equal to its kernel
density calculated from every *other* point. This avoids the issue of obtaining infinite likelihood
at zero bandwidth, which is the case under an ordinary kernel density model.
}
\examples{

# create a data frame of observations
df <- data.frame(x=rnorm(100),y=rnorm(100))

# create a vector of bandwidths to explore
lambda <- seq(0.1,2,0.1)

# obtain log-likelihood at each of these bandwidths
logLike <- kernelLogLike(df,bandwidth=lambda)

# find the maximum-likelihood bandwidth
lambda_ML <- lambda[which.max(logLike)]

# use this value when calculating kernel density distances
distances <- harmonicDist(df,bandwidth=lambda_ML)

}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

