% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Mahalanobis.R
\name{Mahalanobis}
\alias{Mahalanobis}
\title{KE Lotterhos
 Aug 11, 2014
 Updated March 8, 2015
 Empirical p-value Malanahobis distance}
\usage{
Mahalanobis(dfv, column.nums)
}
\arguments{
\item{dfv}{is a dataframe containing the observations in rows and the statistics in columns}

\item{column.nums}{is the column numbers of the dataframe containing the statistics used to calculate Mahalanobis distance}
}
\description{
KE Lotterhos
 Aug 11, 2014
 Updated March 8, 2015
 Empirical p-value Malanahobis distance
}
\author{
KE Lotterhos
}

