% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINOTAUR-package.r, R/MINOTAUR.R
\docType{package}
\name{MINOTAUR}
\alias{MINOTAUR}
\alias{MINOTAUR-package}
\title{MINOTAUR.}
\usage{
MINOTAUR()
}
\description{
MINOTAUR.

Running this function will open the MINOTAUR app, a
shiny application that can be used to explore data, compare variables,
and identify outliers in multivariate and univariate space.
}
\examples{

## run function
MINOTAUR()

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

