% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceFunctions.R
\name{neighborDist}
\alias{neighborDist}
\title{Nearest Neighbor Distance}
\usage{
neighborDist(dfv, column.nums = 1:ncol(dfv))
}
\arguments{
\item{dfv}{a data frame containing observations in rows and statistics in columns.}

\item{column.nums}{indexes the columns of the data frame that contain the statistics used to
calculate nearest neighbor distance (all other columns are ignored).}
}
\description{
Calculates euclidean distance to nearest neighbor for all points.
}
\details{
Takes a matrix or data frame as input, with observations in rows and statistics in columns.
Values are first normalized in each dimension by subtracting from the
mean and dividing by the standard deviation before calculating nearest neighbor distance.

Note that this method cannot handle NA values.
}
\examples{

# create a data frame of observations
df <- data.frame(x=rnorm(100),y=rnorm(100))

# calculate nearest neighbor distances
distances <- neighborDist(df)

# use this distance to look for outliers
Q95 <- quantile(distances, 0.95)
which(distances>Q95)

}
\author{
Robert Verity \email{r.verity@imperial.ac.uk}
}

