% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2D.R
\name{plot_2D}
\alias{plot_2D}
\title{Generate x-y scatter plots.}
\usage{
plot_2D(x, y, xlab, ylab, xlim = NULL, ylim = NULL, n.bins, x_sub, y_sub,
  col.pal, grid = FALSE, outlier.col, outlier.col.bg, outlier.transp,
  outlier.pch, outlier.cex)
}
\description{
Generate x-y scatter plots with outliers overlayed.
Easily control the aesthetics of plot scatter and outlier points.
}
\examples{

}
\author{
Caitlin Collins \email{caitiecollins@gmail.com}
}

