% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TwoRefSim}
\alias{TwoRefSim}
\title{Example Dataset: Simulated Expansion from Two Refugia}
\format{A data frame with 10,000 Single-Nucleotide Polymorphisms (SNPs)
in the rows and 15 variables in the columns.}
\source{
Simulated population genetic dataset.
}
\usage{
TwoRefSim
}
\description{
This dataset contains population genetic data simulated
under a model of expansion from two refugia.
}
\details{
\itemize{
\item \code{SNPnames} \itemize{\item A factor containing the unique names of each SNP locus.}
\item \code{Corr.all} \itemize{\item Correlation between allele frequencies and the environment, based on all 129600 demes on the landscape and infinite sampling of individuals}
\item \code{FST.All} \itemize{\item FST for the SNP based on all 129600 demes on the landscape and infinite sampling of individuals}
\item \code{s_high} \itemize{\item The selection coefficient for the SNP (0 = Neutral)}
\item \code{SNPIncluded} \itemize{\item Was the SNP included in the analysis? (SNPs with He < 0.05 were removed)}
\item \code{He.LS} \itemize{\item Expected heterozygosity (He) for the SNP based on all 129600 demes on the landscape and infinite sampling of individuals}
\item \code{He_samp} \itemize{\item Expected heterozygosity (He) for the SNP based on the sample of finite populations and finite individuals}
\item \code{log.bf} \itemize{\item The log-Bayes Factor from the genetic-environment association output by Bayenv2}
\item \code{rho} \itemize{\item Spearman's rho corrected by population structure output by Bayenv2}
\item \code{xtx} \itemize{\item XTX (an FST analog) output by Bayenv2}
\item \code{TW.Zscore} \itemize{\item Z-score output by LFMM for the genetic-environment association}
\item \code{Md} \itemize{\item Mahalanobis distance based on log.bf, rho, xtx, and TW.Zscore}
\item \code{Hd} \itemize{\item Harmonic mean distance based on log.bf, rho, xtx, and TW.Zscore}
\item \code{Kd} \itemize{\item Kernel density deviance based on log.bf, rho, xtx, and TW.Zscore}
\item \code{Nd} \itemize{\item Nearest neighbor distance based on log.bf, rho, xtx, and TW.Zscore}
}
}
\author{
Kathleen Lotterhos \email{k.lotterhos@neu.edu}

Caitlin Collins \email{caitlin.collins12@imperial.ac.uk}
}
\references{
Lotterhos KE, Whitlock MC (2015)
The relative power of genome scans to detect local adaptation
depends on sampling design and statistical method.
Molecular Ecology, 24, 1031-1046.

Robert Verity*, Caitlin Collins*,
Daren C. Card, Sara M. Schaal, Liuyang Wang, & Katie E. Lotterhos.
MINOTAUR: an R package for visualizing and
calculating multivariate outliers in genomic datasets.
In Review.
}
\keyword{datasets}

